Add addition feature, specific typing and additional testing
