+++ .travis.yml
@@ -17,7 +17,6 @@ deploy:
     secure: APWhf6jTNpcp8YSz5g4H5A/xhEEu5WPbio+yqiHyznADB/3WkI2Z2EeQjTOIVlybgjGj25ZiIFQCvpIUg/meHnISU31utGf9cnsjyit6JlyL3ebjNsc2fiJOnvdQy/hk6N0RHbL5sFYsAmED0YFUJpQsjrJnQwW7QtoMEftrz/FfFrZM0XBiuyZLWgQ3HwpIgGB2Q0Ab1RQhpyFRGSvzovBJ0tbMNWWdjZ2cMDp99UudVQFpm42pUZALztgPk3M2ZRYBFDx6N+aXhm/z7wu/ViX0j+Ji6y1rOm4Oj+zqRn9SsIt78wSPKbQDJfkR7kYyTPPxtc8vpcYq9u/1m+I3XmP4i3CzZzP0XUf6dVuiXrydkVC26+xka2TgxvVVD7jRiIZwGCaN/eqdD4MlHXdeVS9e8IerV5A7yrtkUEo2VPYiqWBGlebzZFkjlv2n0PcVMq7IcrSoyDZQLUZD/R++2HpiJgTCC1SJ4SdfAgF76gOHZI7E9ehAy3C5qa0BC4065RyIgeDd4Z3iWgbK2RjwW6QZrCnB1mzAYnS66TgD0Bw9JaUsBQWx2R2/frFFFy34NDXgWeKYcZGi5Inl6g1LSTmWTxAfP5ZLU+OnuCdMPmihadBzxulKrquRXqauHwQgUWgDaxmElqtfNWQYT4LxV8VA9jyT/zeqiGvek8koCvg=
   on:
     tags: true
-    branch: master
     repo: greenygh0st/ng-timespan
 notifications:
   slack:
