+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ng-timespan",
-  "version": "1.0.1",
+  "version": "1.1.0",
   "description": "A TS/JS implementation of Timespans",
   "main": "ng-timespan.umd.js",
   "jsnext:main": "ng-timespan.esm.js",
